pyproject.toml
[tool.poetry]
name = "python-kafka"
version = "0.1.0"
description = ""
authors = ["GabrielLins64 <gabrielinsmelo@gmail.com>"]
readme = "README.md"
packages = [{include = "python_kafka"}]

[tool.poetry.dependencies]
python = "^3.10"
kafka-python = "^2.0.2"
python-socketio = "^5.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
